Make the shell a registered process
This commit makes the shell a registered process which means
it can receive messages from workers, e.g.:

noedz > send 0 ping
noedz > inbox
(0, 'pong')

This means changing the way we register processes. Instead of
supplying an inbox we now rely on the broker_register fun to
create the queue for us via a Multiprocessing.Manager so that we
can pass it to the broker process.

We also add an inbox function to the shell so we can inspect the
inbox.