Add a function for consistent writes
This commit adds the cput function for consistent key/value writes.

This initial implementation is intentionally naive. It will only
yield consistent writes under perfect conditions and even then
only if the race condition is kind to us (because cput doesn't
even attempt to verify the writes were ok before returning).

The idea here is that the implementation is so bad we should be
able to write numerous test cases that break it. These test cases
then become our benchmark for determining whether we have a cput
implementation that actually works.

Also note there are some assumptions lurking around the code now
(worker pids being integers from 0) and that workers need to know
about each other when they are initialised and cannot be
dynamically added.