Rename multinode/singlenode tests
Rename multinode/singlenode tests to cput/cget and put/get
respectively. This makes it clearer that we are testing single
put/get operations and consistent put/get operations.

We also add a cget function to workers and in keeping with
tradition it is a total fraud - it just returns whatever value
is local to the node. Again, it's up to the tests to expose
this function for the imposter that it is.